refactor: extract grid view into separate file
